Add a World class containing the grid of cells
I wanted to give the program an object oriented nature. I started of
creating a cell with a few important function. Then I realized that it
is actually just a small unit of a bigger world. Therefore, the world
should have information about the cells within it and not the other way
round.

It didn't seem to be a good idea to have each cell know about the whole
world and use that info. to change its status. Rather, the world should
be able to do so - being the 'bigger' entity. A cell should just know
about its position and status.